fix: make ai turns consistent
Found a couple of problems with the previous medium ai and fixed those
in the new hard ai.
The main problem was that the splice function was taking objects out of
the array that was being iterated through at the same
time (avaliableBorders[]). This messed up the iteration after the
splicing.

I fixed this by copying the array and only splicing objects out of the
copy. Important: The array had to be copied like so:
let leftOverBorders = availableBorders.slice();
if the .slice is left out, the leftOverBorders is only a reference to
the original array and not an independent copy.